+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add the `field_path_in_current_schema` property
 * Make sure `alchemy-fields` know which `alchemy-form` they belong to
 * Improve the `original_value` getters
+* Set the `field-type` attribute of `alchemy-field` elements
 
 ## 0.1.3 (2022-02-20)